Move all applcation po's into `application/po` (previously all in `application`)
