+++ src/test-e2e/application/application-create-e2e.spec.ts
@@ -7,8 +7,8 @@ import { e2e } from '../e2e';
 import { ConsoleUserType } from '../helpers/e2e-helpers';
 import { SideNavigation, SideNavMenuItem } from '../po/side-nav.po';
 import { ApplicationE2eHelper } from './application-e2e-helpers';
-import { ApplicationBasePage } from './application-page.po';
-import { CreateApplicationStepper } from './create-application-stepper.po';
+import { ApplicationBasePage } from './po/application-page.po';
+import { CreateApplicationStepper } from './po/create-application-stepper.po';
 
 describe('Application Create', function () {
