+++ src/test-e2e/application/po/application-page-events.po.ts
@@ -1,12 +1,12 @@
-import { ListComponent } from '../po/list.po';
+import { ListComponent } from '../../po/list.po';
 import { ApplicationBasePage } from './application-page.po';
 
-export class ApplicationPageRoutesTab extends ApplicationBasePage {
+export class ApplicationPageEventsTab extends ApplicationBasePage {
 
   list: ListComponent;
 
   constructor(public cfGuid: string, public appGuid: string) {
-    super(cfGuid, appGuid, 'routes');
+    super(cfGuid, appGuid, 'events');
     this.list = new ListComponent();
   }
