+++ src/test-e2e/application/po/application-page-instances.po.ts
@@ -1,11 +1,8 @@
-import { by, element } from 'protractor';
-
-import { ListComponent } from '../po/list.po';
+import { ListComponent } from '../../po/list.po';
 import { ApplicationBasePage } from './application-page.po';
-import { CardAppInstances } from './po/card-app-instances.po';
-import { CardAppStatus } from './po/card-app-status.po';
-import { CardAppUptime } from './po/card-app-uptime.po';
-import { CardAppUsage } from './po/card-app-usage.po';
+import { CardAppInstances } from './card-app-instances.po';
+import { CardAppStatus } from './card-app-status.po';
+import { CardAppUsage } from './card-app-usage.po';
 
 export class ApplicationPageInstancesTab extends ApplicationBasePage {
