+++ src/test-e2e/application/po/application-page-routes.po.ts
@@ -1,12 +1,12 @@
-import { ListComponent } from '../po/list.po';
+import { ListComponent } from '../../po/list.po';
 import { ApplicationBasePage } from './application-page.po';
 
-export class ApplicationPageVariablesTab extends ApplicationBasePage {
+export class ApplicationPageRoutesTab extends ApplicationBasePage {
 
   list: ListComponent;
 
   constructor(public cfGuid: string, public appGuid: string) {
-    super(cfGuid, appGuid, 'variables');
+    super(cfGuid, appGuid, 'routes');
     this.list = new ListComponent();
   }
