+++ src/test-e2e/application/po/application-page-variables.po.ts
@@ -1,12 +1,12 @@
-import { ListComponent } from '../po/list.po';
+import { ListComponent } from '../../po/list.po';
 import { ApplicationBasePage } from './application-page.po';
 
-export class ApplicationPageEventsTab extends ApplicationBasePage {
+export class ApplicationPageVariablesTab extends ApplicationBasePage {
 
   list: ListComponent;
 
   constructor(public cfGuid: string, public appGuid: string) {
-    super(cfGuid, appGuid, 'events');
+    super(cfGuid, appGuid, 'variables');
     this.list = new ListComponent();
   }
