+++ src/test-e2e/application/po/deploy-app.po.ts
@@ -1,7 +1,7 @@
-import { Page } from '../po/page.po';
-import { StepperComponent } from '../po/stepper.po';
-import { TableComponent } from '../po/table.po';
-import { ListComponent } from '../po/list.po';
+import { ListComponent } from '../../po/list.po';
+import { Page } from '../../po/page.po';
+import { StepperComponent } from '../../po/stepper.po';
+import { TableComponent } from '../../po/table.po';
 
 export class DeployApplication extends Page {
