+++ src/test-e2e/applications/applications.po.ts
@@ -1,9 +1,8 @@
-import { protractor, ElementFinder } from 'protractor/built';
-import { browser, element, by, ElementArrayFinder } from 'protractor';
+import { by, element } from 'protractor';
+
+import { DeployApplication } from '../application/po/deploy-app.po';
 import { CFPage } from '../po/cf-page.po';
 import { ListComponent } from '../po/list.po';
-import { id } from '@swimlane/ngx-charts/release/utils';
-import { DeployApplication } from '../application/deploy-app.po';
 
 export class ApplicationsPage extends CFPage {
